fix: remove integer constraint from all remaining NGN money fields
Audit found the same integer-for-money anti-pattern in 7 more places
that were missed by migration 0048. The root cause was that migration
0039 established the pattern of using 'integer' for naira columns, and
every subsequent migration copied it. The Zod schemas and forms then
mirrored the database constraint.

Database (migration 0049):
- commerce_events.price_ngn: integer -> numeric(12,2)
- community_observations.amount_ngn: integer -> numeric(12,2)

Zod schemas (removed .int() from money fields):
- lib/analytics/commerce-events.ts: priceNgn
- lib/community-intake/schema.ts: priceNgn, amountNgn
- lib/retailer-partnership/schema.ts: samplePriceNgn (2 places)
- lib/campaigns/daily-desk.ts: priceNgn
- lib/moderation/retailer-presentation.ts: ngnAmount

UI inputs (inputMode='numeric' -> 'decimal', added step='0.01'):
- components/contribute/contribution-experience.tsx: price input
- components/retailers/retailer-partnership-experience.tsx: sample price

Model validation (Number.isInteger -> Number.isFinite):
- lib/commerce/assisted-procurement-model.ts: quoteTotal()
- modules/commerce/verified-retail-offers.test.ts: price assertion

The offers.price_minor and offer_price_history.price_minor columns
are correctly using bigint for minor units (kobo) and were not changed.

+++ lib/campaigns/daily-desk.ts
@@ -25,7 +25,7 @@ const acceptedCampaignSchema = z.object({
   offerEvidence: z
     .array(
       z.object({
-        priceNgn: z.number().int().positive().max(100_000_000),
+        priceNgn: z.number().positive().max(100_000_000),
       }),
     )
     .min(1)
